refactor: PR 요구사항 반영
