+++ app/src/main/java/com/example/swap_it/data/datasource/local/model/bottomsheet/CategoryOption.kt
@@ -1,14 +1,17 @@
 package com.example.swap_it.data.datasource.local.model.bottomsheet
 
-enum class CategoryOption(val categoryName: String) {
-    CAR("자동차"),
-    CLOTH("옷"),
-    MOBILE("모바일 기기"),
-    CUP("컵"),
-    BOOK("책"),
-    ELECTRONIC("가전기기"),
-    FASHION("패션"),
-    FOOD("음식"),
-    ETC("기타"),
-    SNACK("과자"),
+enum class CategoryOption(val option: String) {
+    ELECTRONICS("전자기기"),
+    APPLIANCES("가전제품"),
+    FURNITURE("가구"),
+    CLOTHING("의류/패션"),
+    BEAUTY("뷰티/화장품"),
+    SPORTS("스포츠/레저"),
+    TOYS("장난감/키즈"),
+    BOOKS("책/문구"),
+    AUTOMOTIVE("자동차/부품"),
+    PET("반려동물용품"),
+    GROCERIES("식품/음료"),
+    HOBBY("취미/게임/음반"),
+    MISC("기타"),
 }
