+++ app/src/main/java/com/example/swap_it/ui/alert/AlertScreen.kt
@@ -16,7 +16,6 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.example.swap_it.R
 import com.example.swap_it.ui.component.BackButton
-import com.example.swap_it.ui.component.Cards
 import com.example.swap_it.ui.theme.BackgroundColor
 import com.example.swap_it.ui.theme.Paddings
 import com.example.swap_it.ui.theme.SwapitTheme
@@ -39,7 +38,7 @@ fun AlertScreen(modifier: Modifier = Modifier, navController: NavHostController)
         )
         LazyColumn(horizontalAlignment = Alignment.CenterHorizontally) {
             items(4) {
-                Cards().AlertCard(alertCardData = Cards.alertCardData)
+                AlertCard(alertCardData = alertCardData)
             }
         }
     }
