+++ app/src/main/java/com/example/swap_it/ui/shopping/ModalBottomSheetContent.kt
@@ -104,7 +104,7 @@ fun CategoryButtons(option: SortOption) {
         ) {
         CategoryOption.entries.forEach {
             CategoryButton(
-                text = it.categoryName,
+                text = it.option,
                 isSelected = true,
                 modifier = Modifier.padding(Paddings.small)
             ) {
