+++ app/src/main/java/com/example/swap_it/ui/swap/RequestedSwapSection.kt
@@ -18,7 +18,6 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.example.swap_it.R
-import com.example.swap_it.ui.component.Cards
 import com.example.swap_it.ui.theme.Paddings
 import com.example.swap_it.ui.theme.Typography
 
@@ -50,7 +49,7 @@ import com.example.swap_it.ui.theme.Typography
         }
         LazyRow {
             items(10){
-                Cards().SwapCard(swapCardData = Cards.swapCardData)
+                SwapCard(swapCardData = swapCardData)
             }
         }
     }
